--- src/Header.js
@@ -1,13 +1,0 @@
-const Header = () => {
-    return (
-        <header>
-            <div className="wrapper">
-                <h1>Everything Music</h1>
-                <p>Welcome to your one stop shop for information about any and all songs</p>
-                <p>Enter a song title and artist below, and we will dig into our database and get you some facts!</p>
-            </div>
-        </header>
-    )
-}
-
-export default Header;
